feat: add editor num setting button
